✨ Add deep modeling process execution script and utilities
- Add executeDeepModelingProcess.ts script for comprehensive database design workflow
- Create shared utilities (scriptUtils.ts, streamingUtils.ts) for script functionality
- Update agent package.json with new script dependency
- Enhance executeDesignProcess.ts with improved workflow handling
- Add comprehensive business management system design capabilities
