+++ frontend/internal-packages/agent/package.json
@@ -32,6 +32,9 @@
   },
   "scripts": {
     "dev": "tsc --watch",
+    "execute-deep-modeling": "tsx scripts/executeDeepModelingProcess.ts",
+    "execute-deep-modeling:debug": "tsx scripts/executeDeepModelingProcess.ts --log-level=DEBUG",
+    "execute-deep-modeling:warn": "tsx scripts/executeDeepModelingProcess.ts --log-level=WARN",
     "execute-design-process": "tsx scripts/executeDesignProcess.ts",
     "execute-design-process:debug": "tsx scripts/executeDesignProcess.ts --log-level=DEBUG",
     "execute-design-process:warn": "tsx scripts/executeDesignProcess.ts --log-level=WARN",
